# feat: more encoding, errors, spec tests