--- a/src/base.js
+++ b/src/base.js
@@ -4,22 +4,22 @@ class Base {
   constructor (name, code, implementation, alphabet) {
     this.name = name
     this.code = code
+    this.codeBuf = Buffer.from(this.code)
     this.alphabet = alphabet
-    if (implementation && alphabet) {
-      this.engine = implementation(alphabet)
+    this.engine = implementation(alphabet)
+  }
+
+  encode (buf) {
+    return this.engine.encode(buf)
+  }
+
+  decode (string) {
+    for (const char of string) {
+      if (this.alphabet && this.alphabet.indexOf(char) < 0) {
+        throw new Error(`invalid ${this.name} character '${char}' in '${string}'`)
+      }
     }
-  }
-
-  encode (stringOrBuffer) {
-    return this.engine.encode(stringOrBuffer)
-  }
-
-  decode (stringOrBuffer) {
-    return this.engine.decode(stringOrBuffer)
-  }
-
-  isImplemented () {
-    return this.engine
+    return this.engine.decode(string)
   }
 }