--- a/test/constants.spec.js
+++ b/test/constants.spec.js
@@ -10,11 +10,11 @@ const constants = require('../src/constants.js')
 describe('constants', () => {
   it('constants indexed by name', () => {
     const names = constants.names
-    expect(Object.keys(names).length).to.equal(16)
+    expect(Object.keys(names).length).to.equal(23)
   })
 
   it('constants indexed by code', () => {
     const codes = constants.codes
-    expect(Object.keys(codes).length).to.equal(16)
+    expect(Object.keys(codes).length).to.equal(23)
   })
 })