… .travis.yml
… README.md
… package.json
… src/base16.js
… src/base32.js
… src/base64.js
… test/spec-test5.spec.js
… test/spec-test6.spec.js