@fabianhjr 2020-06-09
Hi, multiformats/multibase#60 has been merged so the changed test vectors should now match with master.